refactor: 금액을 나타내는 타입 BigDecimal로 변경

+++ src/main/java/roomescape/service/request/PaymentApproveAppRequest.java
@@ -1,8 +1,9 @@
 package roomescape.service.request;
 
+import java.math.BigDecimal;
 import roomescape.web.controller.request.MemberReservationRequest;
 
-public record PaymentApproveAppRequest(String paymentKey, String orderId, Long totalAmount) {
+public record PaymentApproveAppRequest(String paymentKey, String orderId, BigDecimal totalAmount) {
 
     public static PaymentApproveAppRequest from(MemberReservationRequest request) {
         return new PaymentApproveAppRequest(request.paymentKey(), request.orderId(), request.amount());
